fix(build): satisfy GCC -Werror=useless-cast and -Werror=missing-field-initializers
splitmix64(sym) & mask already evaluates to uint64_t which is size_t on 64-bit targets, so the outer cast in shard_index_for is redundant under -Werror=useless-cast.

gen_state designated-initializer braces with .rng = ... left .next_id and .live uninitialized, which GCC flags as -Werror=missing-field-initializers. Initialize all three fields explicitly at the three call sites in test_engine_differential.cpp.

+++ include/lob/shard_router.hpp
@@ -67,7 +67,7 @@ class shard_router {
     }
 
     [[nodiscard]] std::size_t shard_index_for(symbol_id_t sym) const noexcept {
-        return static_cast<std::size_t>(splitmix64(sym) & mask);
+        return splitmix64(sym) & mask;
     }
 
     [[nodiscard]] static constexpr std::size_t shard_count() noexcept { return NumShards; }
